+++ spynnaker/pyNN/models/neural_projections/connectors/fixed_number_pre_connector.py
@@ -213,7 +213,7 @@ class FixedNumberPreConnector(AbstractGenerateConnectorOnMachine,
             n_connections_total, prob_in_slice, chance=1.0/100000.0)
 
         if min_delay is None or max_delay is None:
-            return int(math.ceil(n_connections))
+            return math.ceil(n_connections)
 
         return self._get_n_connections_from_pre_vertex_with_delay_maximum(
             synapse_info.delays,
